+++ db/heap/heap_alloc_job.cc
@@ -85,10 +85,10 @@ Status HeapAllocJob::Add(const Slice& key, const Slice& value,
   ctx_.allocated_ = true;
 
   uint32_t checksum = Lower32of64(XXH3_64bits(value.data_, value.size()));
-  *hvi = HeapValueIndex(
-      ikey.sequence, ctx_.current_ext_id_, static_cast<uint16_t>(offset),
-      static_cast<uint16_t>(need_block), static_cast<uint32_t>(value.size()),
-      checksum, kNoCompression);
+  *hvi = HeapValueIndex(ikey.sequence, ctx_.current_ext_id_,
+                        static_cast<uint32_t>(offset), need_block,
+                        static_cast<uint32_t>(value.size()), checksum,
+                        kNoCompression);
   if (UNLIKELY(value.size() > kBufferSize)) {
     SubmitValueInBuffer();
     void* ptr = std::aligned_alloc(4096, aligned_value_size);
@@ -113,7 +113,7 @@ Status HeapAllocJob::Add(const Slice& key, const Slice& value,
     return s;
   }
 
-  if (ctx_.base_bno_ + ctx_.cnt_ == offset) {
+  if (ctx_.base_bno_ + ctx_.cnt_ == static_cast<uint32_t>(offset)) {
     // seq write
     ctx_.cnt_ += need_block;
   } else {
